+++ include/server.h
@@ -20,7 +20,7 @@ void childClose();
 void handleSIGCHLD(int signum); 
 
 /* Client command paring */
-void decode_command(client_t *client, char *command, int *channel_id, char *message);
+void decode_command(client_t *client, char *command, long *channel_id, char *message);
 
 /* Setup a new client object after connection */
 void client_setup(client_t *client, int client_id);
@@ -32,22 +32,22 @@ void client_processing(client_t *client, pid_t current_process);
 // Action methods //
 
 /* Subscribe CLIENT to CHANNEL_ID */
-void subscribe(int channel_id, client_t *client);
+void subscribe(long channel_id, client_t *client);
 
 /* List all channels CLIENT is connect to, and info about those channels */
 void channels(client_t *client);
 
 /* Unsubscribe CLIENT from CHANNEL */
-void unsubscribe(int channel_id, client_t *client);
+void unsubscribe(long channel_id, client_t *client);
 
 /* Show next unread message from any channel CLIENT is connected to */
 void next(client_t *client);
 
 /* Show next unread message from CHANNEL_ID if CLIENT is connected to it */
-void nextChannel(int channel_id, client_t *client);
+void nextChannel(long channel_id, client_t *client);
 
 /* Send MSG to CHANNEL from CLIENT */
-void sendMsg(int channel_id, client_t *client, char *message);
+void sendMsg(long channel_id, client_t *client, char *message);
 
 /* Disconnect CLIENT from the server */
 int bye(client_t *client);
@@ -59,10 +59,10 @@ int bye(client_t *client);
 msgnode_t* node_add(msgnode_t *head, msg_t *message);
 
 /* Read message in specific channel and move the client read head */
-msg_t* read_message(int channel_id, client_t *client);
+msg_t* read_message(long channel_id, client_t *client);
 
 /* Just retrieve next message in specific channel but do NOT move client read head */
-msg_t* get_next_message(int channel_id, client_t *client);
+msg_t* get_next_message(long channel_id, client_t *client);
 
 /* Retrieve the number of unread messages */
-int get_number_unread_messages(int channel_id, client_t *client);
+int get_number_unread_messages(long channel_id, client_t *client);
